formatter: Also allow /*...*/ as inline comments
Blueprint allows both `//` and `/*...*/` style comments, but if a
project prefers only `/*...*/` comments, it is currently not possible to
have these inline.

Therefore, treat these comments equal if they occur inline. To make this
easier to understand, we refactor the comment handling slightly to first
handle single-line comment whitespace, and then handle newlines for both
single-line and multi-line style comments.

Adjust the test accordingly to make sure this works.