decompiler: Force newline before translator comments
The decompiler does not insert newlines and so relies on the formatter
to format its output. The previous commit however broke translator
comments, as they are now assumed to be inline.

Therefore, make sure the decompiler forces a newline before a translator
comment. We could special case such a comment in the formatter, but this
might be difficult if #202 gets resolved, whereas in the decompiler we
already know exactly when a comment is a translator comment.