scale: Decompile non-translatable labels too
While working on the previous commit, I noticed Gtk.Scale specially
handles translator comments, but more importantly, that it does not put
non-translatable labels in its decompiled output.

This is wrong, as it's not unlikely that Gtk.Scale marks would use a
purely numeric string, which should not be lost upon decompilation.

Add a test for this case too.